Nasm file deletion workaround
Nasm will delete the output file on invalid asm. Normally this file is
/dev/stdout, which is a symlink to /proc/self/fd/1, and nasm will fail
with EPERM when attempting to delete things in /dev. Howevr if run as
root, this op succeeds and the symlink is nuked.

Change the path to /proc/self/fd/1, meaning even if nasm fails and
deletes the file, it will only delete it's own fd and leave /dev
untouched.

Fixes #14, thanks to @thatreguy for the bug report.